Added collision on X's and O'x to the players movements
Feeling a bit lazy so figured that instead of storing the prevoius
positions state, the player just can't move there. This decision has
cost me more than I'd like to admit as far as time goes to make
possible, and I can already think of a case where this is screwed.

 x
oPo
 x

If P is the player cursor, then they are stuck, as they cannot jump over
anywhere, yet that still needs to be a valid place to place a token. In
this situation I'll need to design more code, perhaps moving down or up
places you in the next avaliable square in that row? Doing that would
also fix the following problem

Pxo
x
x

Here the player can also not escape