Refined movement, including collision rules
Fixed the movement so that at all times the player can get to any empty
squares on the board. I hope the movement is intuitve.

I have commmented this code a lot more than I usually would, but I think
the way it is currently written it is quite confusing.