Computer now places randomly every turn
This one took embarrasingly long to figure out, even made a unit test it
was taking so long. Fixed now though. Now whenever the player places a
token, the computer does too. At the moment there is no semberlance of
strategy, it just places it. Next need to be able to declare a winner